Use Compose State instead of MutableStateFlow and Flow.combine.
We are moving towards a factorisation where the ViewModel contains state and coroutines that are supposed to survive activity recreation. Other logic though, like the flow collection, is only meant to happen when the UI exists, and this doesn't belong in the ViewModel. I don't have a good name for that new class though. I am tempted to say that the logic that runs only when the UI is there is in fact @Composable UI logic.

 Conflicts:
	app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksScreen.kt
	app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksViewModel.kt