Generate TasksUiState using an @Composable function, instead of Flow.combine().
This is the "transpose" of the code in the parent commit. Here the logic is arranged by the type of AsyncTask. (In the parent commit, we aranged the logic by the UI state property in question.)

Advantage: we only need to collect tasksAsyncFlow once.

Disadvantage: this code is harder to split up into N pieces.